+++ examples/basic middleware/index.d.ts
@@ -1,7 +1,0 @@
-import { HTTPRequestContext } from "rjweb-server"
-
-declare module "my-middleware" {
-  export interface Props {
-    printEmpty: () => HTTPRequestContext
-  }
-}