+++ templates/[Javascript] Basic JSON API/src/routes/echo.js
@@ -1,0 +1,9 @@
+import { RouteFile } from "rjweb-server"
+
+export default new RouteFile((file) => file
+  .http('POST', '/echo', (http) => http
+    .onRequest((ctr) => {
+      ctr.print({ content: ctr.rawBody, requests: ctr["@"].requests })
+    })
+  )
+)