+++ templates/[Javascript] Basic JSON API/src/routes/hello.js
@@ -1,0 +1,9 @@
+import { RouteFile } from "rjweb-server"
+
+export default new RouteFile((file) => file
+  .http('GET', '/hello', (http) => http
+    .onRequest((ctr) => {
+      ctr.print({ hello: 'world', requests: ctr["@"].requests })
+    })
+  )
+)