+++ templates/[Typescript] Basic JSON API/src/routes/echo.ts
@@ -1,0 +1,10 @@
+import { RouteFile } from "rjweb-server"
+import { WebServerContext } from "../types/context"
+
+export = new RouteFile((file) => file
+  .http<WebServerContext>('POST', '/echo', (http) => http
+    .onRequest((ctr) => {
+      ctr.print({ content: ctr.rawBody, requests: ctr["@"].requests })
+    })
+  )
+)