+++ templates/[Typescript] Basic JSON API/src/routes/hello.ts
@@ -1,0 +1,10 @@
+import { RouteFile } from "rjweb-server"
+import { WebServerContext } from "../types/context"
+
+export = new RouteFile((file) => file
+  .http<WebServerContext>('GET', '/hello', (http) => http
+    .onRequest((ctr) => {
+      ctr.print({ hello: 'world', requests: ctr["@"].requests })
+    })
+  )
+)